chore: move FB test harness to flashblocks crate
